+++ packages/expo-router/src/modal/Modal.tsx
@@ -23,7 +23,7 @@ export interface ModalProps extends ViewProps {
   visible: boolean;
   /**
    * Callback that is called after modal is closed.
-   * This is called when the modal is dismissed by the user or programmatically.
+   * This is called when the modal is closed programmatically or when the user dismisses it.
    */
   onClose?: () => void;
   /**
@@ -102,7 +102,7 @@ export function Modal(props: ModalProps) {
     transparent,
     ...viewProps
   } = props;
-  const { openModal, closeModal, addEventListener } = useModalContext();
+  const { openModal, updateModal, closeModal, addEventListener } = useModalContext();
   const [currentModalId, setCurrentModalId] = useState<string | undefined>();
   const navigation = useNavigation<NavigationProp<ParamListBase>>();
   useEffect(() => {
@@ -111,7 +111,7 @@ export function Modal(props: ModalProps) {
     }
   }, [props.detents]);
   useEffect(() => {
-    if (!currentModalId && visible) {
+    if (visible) {
       const newId = nanoid();
       openModal({
         animationType,
@@ -119,20 +119,25 @@ export function Modal(props: ModalProps) {
         transparent,
         viewProps,
         component: children,
-        detents: props.detents,
         uniqueId: newId,
         parentNavigationProp: navigation,
+        detents: props.detents,
       });
       setCurrentModalId(newId);
       return () => {
         closeModal(newId);
       };
-    } else if (currentModalId && !visible) {
-      closeModal(currentModalId);
-      setCurrentModalId(undefined);
     }
     return () => {};
   }, [visible]);
+
+  useEffect(() => {
+    if (currentModalId && visible) {
+      updateModal(currentModalId, {
+        component: children,
+      });
+    }
+  }, [children]);
 
   useEffect(() => {
     if (currentModalId) {
@@ -153,6 +158,6 @@ export function Modal(props: ModalProps) {
       };
     }
     return () => {};
-  }, [currentModalId, addEventListener, onClose]);
+  }, [currentModalId, addEventListener, onClose, onShow]);
   return null;
 }